requirements.txt
# ── Core data ────────────────────────────────────────────────────────────────
pandas>=2.1.0
numpy>=1.26.0
polars>=0.20.0
pyarrow>=14.0.0

# ── HTTP & API ────────────────────────────────────────────────────────────────
requests>=2.31.0
python-dotenv>=1.0.0

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.2.0
optuna>=3.5.0
shap>=0.44.0
statsmodels>=0.14.0

# ── MLOps ────────────────────────────────────────────────────────────────────
mlflow>=2.10.0
evidently>=0.4.20

# ── Portfolio / Finance ───────────────────────────────────────────────────────
PyPortfolioOpt>=1.5.5

# ── API serving ───────────────────────────────────────────────────────────────
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# ── Dashboard ────────────────────────────────────────────────────────────────
streamlit>=1.31.0
plotly>=5.19.0
streamlit-extras>=0.4.0
streamlit-lottie>=0.0.5
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0

# ── Code quality ─────────────────────────────────────────────────────────────
flake8>=7.0.0
black>=24.0.0
